Library mindstorm

OCaml Mindstorm is a library that enables to drive LEGO® mindstorm robots from OCaml through a Bluetooth or USB connection.

Mindstorm — Drive Lego Minsdstorm bricks with OCaml.

How to connect to the Lego brick is explained at the end of the API description, in the sections

"connectBluetooth" and

"connectUSB".